rust_patterns::data_structures

Module value_vector

ValueVec<T>: a by-value, ref-less vector with interior mutability.

You can convert to and from a Vec<T> at zero cost.

Key properties:

All mutating operations use &self (immutable receiver).

No references to elements are ever given out.

Elements are inserted/removed/moved by value.

Getting a value returns a Clone (or Copy) of the stored element.

Many shared immutable references to a ValueVec can exist simultaneously safely.

Trade-offs:

You can’t borrow into the backing memory; you pay cloning cost to read.

In return, the backing storage may reallocate at will without invalidating any external references (because none are ever issued).

Functionality:

For any functionality of Vec<T> that ValueVec<T> doesn’t provide, you can use into_vec to convert to a Vec<T> at zero cost.

Safety and Soundness:

The fundamental requirement for soundness is that a mutable borrow of the internal Vec (or its owned memory) can never occur simultaneously with any other borrow. The only mutable borrows that exist are the temporary borrows in the method implementations we control. Therefore, ValueVec methods must not be re-entered while another method is still executing on the same instance. Doing so would create overlapping borrows of the internal Vec and cause undefined behavior. Our implementation makes no such explicit calls, so the only way this can happen is implicitly, in particular when the compiler inserts a call to the V type’s Drop implementation. You would have to try pretty hard to trigger this, but it is possible.

There are a few ways to deal with this requirement:

Require V to be Copy. Copy types are not allowed to have destructors. This is the simplest and easiest solution, but it limits ValueVec<V> to a far more restrictive category of types than is necessary.

It might be possible to carefully restrict the API of ValueVec and the implementation of its methods so that re-entrancy does not cause UB, but this is very delicate and an open question.

Enforce a documented contract that V must not require destruction by wrapping all V values in std::mem::ManuallyDrop<V>, which will prevent any destructor from running.

Prevent creation of a ValueVec<V> with an assertion of the form assert!(!needs_drop::<T>(), "Type must not have a destructor"). (Perhaps used in combination with the strategy of the previous bullet.)

Just make users pinky promise not to do weird things in the Drop implementation of V. This dispenses with machine-checked correctness in favor of a human-checked contract, which may be unpalatable or even prohibited in some settings, but which is probably good enough for the vast majority of use cases.

Structs

ValueVec

A by-value, ref-less vector with interior mutability. Values of type V can be moved into and out of the vector. Cloning / copying getters exist if V implements Clone/Copy. See the module-level docs for important notes about soundness.