rust_patterns::plugins

Module item_registry

The “Registry Pattern”

The “Registry Pattern” is a plugin-like pattern for managing a collection of RegisteredItem instances such that discovery of items is automatic/implicit.

For our use case, the RegisteredItems are lazily registered (the allocation or instantiation of static data) and instantiated (for instance data), but this is not an essential part of the pattern. What is essential is that the RegisteredItems are automatically discovered, even if they are defined in external crates.

In the simplest formulation, the “registry pattern” is just a vector of RegisteredItems each one of which has a static index variable in which is stored its position in the vector that owns it. We just provide a mechanism for assigning the indices to the RegisteredItems at runtime in a logically consistent and thread-safe way that doesn’t require explicit knowledge of every RegisteredItem type that exists.

We define the RegisteredItems struct as a wrapper type around the vector of RegisteredItem instances to provide convenient RegisteredItems::get_item and RegisteredItems::get_item_mut methods, which automatically instantiate the RegisteredItem instance if it has not yet been initialized. Note that we intentionally do not call the RegisteredItems struct a “registry” because it is not necessarily a singleton. In fact, we might have many RegisteredItems instances each holding RegisteredItem instances it owns. “The registry” that records which RegisteredItem types exist is purely abstract; depending on the underlying implementation, there might not be any actual list of which RegisteredItem types exist at all!

A RegisteredItem need not actually be a singleton, but it, or rather its type, is only ever registered once at runtime; that is, its index, which is shared among all instances of the type, is initialized only once. This allows a “registered item” type to have “singleton metadata” ¹ associated with it in addition to whatever instance data it needs. As already mentioned, at minimum each RegisteredItem knows its own index, which is used to access the item from the owning RegisteredItems container via simple indexing into a vector of RegisteredItems. (See the ‘plugins’ module for more details about implementation.)

Q: Wait, aren’t you just talking about static variables? Or, aren’t you just talking about lazy initialization of static data?

A: No, I am describing an interface for automatic discovery of special types and, (though only incidentally) the initialization of their static data. How rich the static data is depends entirely on your use case, so long as it at minimum has an index field to make accessing instances of the items in the owning container easy.

Notes

1. This is usually just called static data, or data in a static variable, but I am trying to avoid any assumptions about when this data is initialized and whether it is mutable.

Structs

RegisteredItems

Statics

NEXT_ITEM_INDEX

Global item index counter, keeps track of the index that will be assigned to the next item that requests an index. Equivalently, holds a count of the number of items currently registered.

REGISTERED_ITEMS

For simple applications, just knowing how many items are registered is enough, and that information is already captured in NEXT_ITEM_INDEX. In more sophisticated implementations, we would want to store metadata about the items that are registered. For example, we might want to track dependencies between registered items. We could store that metadata in a static variable, or we could have a global data store like the following.

Traits

RegisteredItem

A trait for items that can be registered (DataPlugin, PersonProperty)

Functions

add_to_registry

Adds a new item to the registry. The job of this method is to create whatever “singleton” data/metadata is associated with the RegisteredItem if it doesn’t ready exist. The “registry” is a global singleton (literally just NEXT_ITEM_INDEX in this implementation), so it is enough to know the concrete type of the RegisteredItem. For this illustration, the only singleton metadata that is instantiated is the item’s index, which we return in case the caller wants it.

get_data_plugin_count

initialize_item_index

Acquires a global lock on the next available plugin index, but only increments it if we successfully initialize the provided index. The index of a registered item is assigned at runtime but only once per type. It’s possible for a single type to attempt to initialize its index multiple times from different threads, which is why all this synchronization is required. However, the overhead is negligible, as this initialization only happens once upon first access.

Type Aliases

BxRegisteredItem

A boxed RegisteredItem