rust_patterns::plugins::item_registry

Trait RegisteredItem

pub trait RegisteredItem: Any {
    // Required methods
    fn name() -> &'static str
       where Self: Sized;
    fn index() -> usize
       where Self: Sized;
    fn new_boxed() -> Box<Self>
       where Self: Sized;
}

A trait for items that can be registered (DataPlugin, PersonProperty)

Required Methods

fn name() -> &'static str
where Self: Sized,

Convenient for debugging.

fn index() -> usize
where Self: Sized,

The index of the item in the RegisteredItem list, typically just a vector that lives in the parent.

fn new_boxed() -> Box<Self>
where Self: Sized,

Creates a new instance of the item.