rust_patterns::shared_implementation

Module sealed_traits

Fine-grained Encapsulation and Information Hiding with Traits

This module demonstrates fine-grained access/overridability control of trait methods as described by Predrag Gruevski in his excellent article A definitive guide to sealed traits in Rust. We demonstrate each of the following:

all methods callable downstream	some methods callable downstream	no methods callable downstream

all methods overridable	✅ (pub trait)	❌	❌

some methods overridable	✅ (signature-sealed default methods + pub fn to call them / “final methods” pre-RFC)	✅ (partially-sealed / “final methods” pre-RFC)	❌

trait cannot be impl’d downstream (no methods overridable)	✅ (supertrait sealed)	✅ (at least one signature-sealed method with no default impl)	✅ (all methods signature-sealed)

These techniques enable not just shared interfaces but shared implementation while maintaining a level of encapsulation.

Information Hiding and Encapsulation in Rust

Rust’s language features enabling information hiding and encapsulation are much simpler and fewer in number than those in C++. In Rust, items and fields are private by default, and the visibility modifier pub is intentionally coarse-grained, limited to pub (without qualification), pub(crate), pub(super), and pub(in path). There is no mechanism to grant selective access to private items. Rust has no concept of inheritance. Rust uses traits, which define a set of method signatures, to define shared interfaces. A trait can have one or more trait constraints*, which look like superclasses if you squint, but which are really just assertions that an implementor of the trait must also implement some other trait.

Rust’s support for shared implementation is very limited. A trait can have a default implementation for a method, but even this support for shared implementation is hamstrung by the fact that traits cannot have data members of any visibility. Thus any default implementation can only assume about self what the trait itself asserts and cannot access any data directly. What’s more, as a mechanism for defining public interfaces, trait methods are necessarily public.

However, it turns out that it nonetheless is possible to achieve very fine-grained access and overridability restrictions on traits using a concept called sealed traits.

Modules

all_callable_some_overridable

no_restrictions

Structs

Person

A struct we might want to implement a trait for.