rust_patterns

Module type_erased_api

Typed and Type-Erased API

An illustration of a type, a database index, having both a typed and type-erased API.

Sometimes an implementation needs access to a concrete type while providing a type-erased interface. In some situations, a type-erased interface can provide the same functionality as a typed-API but at some cost, and so you want to provide both a typed and type-erased interface.

In this context, by typed I mean that a concrete type is known at compile time. Generally a typed API will be generic over the type but monomorphized at compile time. (Stop worrying about how monomorphization increases the size of the final binary. And, while you’re at it, stop obsessing about string copies.) The concrete type is Index<T> (notice the generic T). The typed API is just the methods provided by impl<T> Index<T>{...}.

A type-erased API, on the other hand, will be monomorphic over a trait object. We define the type-erased API via the trait TypeErasedIndex (notice the lack of a generic T) and provide a blanket implementation for all Index<T>s.

Notice that it turns out to be convenient to implement many of the methods of the typed API by deferring to the type-erased API, which is typical. You will want to do this when you can to keep your code DRY. Don’t be a slave to the DRY principle; violate it whenever it makes sense. These rules of thumb are not an excuse to turn off your brain.

Structs

Index

The typed Index<T>

Traits

TypeErasedIndex

This trait Encapsulates the type-erased API.

Type Aliases

BxIndex

A “boxed” TypeErasedIndex, use anywhere you need a type-erased Index<T>

EntityId

HashValue