rust_patterns::type_erased_api

Struct Index

pub struct Index<T: Hash + Eq + Clone + Any> {
    lookup: HashTable<(T, HashSet<EntityId>)>,
}

The typed Index<T>

Fields
lookup: HashTable<(T, HashSet<EntityId>)>

We store a copy of the value here so that we can iterate over it in the typed API, and so that the type-erased API can access some serialization of it.

Implementations

impl<T: Hash + Eq + Clone + Any> Index<T>

Contains the typed API

pub fn new() -> Self

pub fn insert_entity(&mut self, key: &T, entity_id: EntityId) -> bool

Inserts an entity into the set associated with key, creating a new set if one does not yet exist. Returns a bool according to whether the entity_id already existed in the set. Observe that several of these just defer to the untyped implementation.

pub fn insert_value( &mut self, key: T, set: HashSet<EntityId>, ) -> OccupiedEntry<'_, (T, HashSet<EntityId>)>

Inserting a new value requires the value itself.

pub fn get(&self, key: &T) -> Option<&HashSet<EntityId>>

Gets an immutable reference to the set associated with the key if it exists. Observe that we just defer to the untyped implementation.

pub fn get_mut(&mut self, key: &T) -> Option<&mut HashSet<EntityId>>

Gets a mutable reference to the set associated with the key if it exists. Observe that we just defer to