rust_patterns::type_erased_api

Trait TypeErasedIndex

pub trait TypeErasedIndex {
    // Required methods
    fn insert_entity_with_hash(
        &mut self,
        hash: HashValue,
        entity_id: EntityId,
    ) -> Result<bool, ()>;
    fn get_with_hash(&self, hash: HashValue) -> Option<&HashSet<EntityId>>;
    fn get_with_hash_mut(
        &mut self,
        hash: HashValue,
    ) -> Option<&mut HashSet<EntityId>>;
    fn has_hash(&self, hash: HashValue) -> bool;
}

This trait Encapsulates the type-erased API.

Required Methods

fn insert_entity_with_hash( &mut self, hash: HashValue, entity_id: EntityId, ) -> Result<bool, ()>

Inserting a new entity only requires the hash but requires the set associated with the hash to already exist.

If the set corresponding to the hash exists, inserts the entity_id into the associated set, returning a bool according to whether the entity_id was already in the set. If the set does not exist, returns Err(())

fn get_with_hash(&self, hash: HashValue) -> Option<&HashSet<EntityId>>

Fetching a set only requires the hash.

fn get_with_hash_mut( &mut self, hash: HashValue, ) -> Option<&mut HashSet<EntityId>>

Fetching a set only requires the hash.

fn has_hash(&self, hash: HashValue) -> bool

Does the index contain the given hash?