rust_patterns::type_erasure

Module static_interface

ZST-based interface to static data of a non-ZST type

Let’s design a ZST-based interface to the static methods/data of a (possibly) non-ZST type (in Rust). The idea is that the concrete implementor M of some trait exposes its static data (or methods) through its implementation of a trait, which we’ll call MyTrait. Suppose we want to interact with that data without having an instance of the concrete type M.

Our goal is to create a type the instances of which provide type-erased proxy access to the static data/methods of interest.

simple_static_vtable shows how to do this using a simple static VTable: Type erasure is achieved by storing a function pointer.

static_reference_to_zst shows how to do this using a static reference to a ZST: Type erasure is achieved by using a zero-sized trait object behind a reference, which does not allocate.

Both strategies use the same technique of generating static constant data

Modules

simple_static_vtable

Using a simple static VTable

static_reference_to_zst

Using a static reference to a ZST

Structs

MyStaticData

The static data we want to expose.

Traits

MyTrait

Our example trait implemented by some concrete type M provides an interface to static methods or data.