rust_patterns::type_erasure::static_interface::simple_static_vtable

Function tag_for

pub fn tag_for<M: MyTrait>() -> Tag

Factory: return a 'static reference to the per-M vtable. Compare to the equivalent function in static_reference_to_zst, which has an additional 'static trait constraint on M due to the returned Tag type being a trait object rather than a concrete static type.