rust_patterns::type_erasure::static_interface

Module static_reference_to_zst

Using a static reference to a ZST

As in the strategy of simple_static_vtable, a single static instance for each concrete type M is stored as an associated constant within a trait having M as a generic parameter. However, type erasure is achieved by storing a reference to a ZST instead of a function pointer. The “token” value is two machine words instead of one.

First, we create a ZST TypedTag<M: MyTrait> to which we can attach the concrete implementing type. Its “job” is really just to be able to “name” the concrete type. We can create (zero-sized) instances of this new type without creating instances of the concrete (potentially non-zero-sized) type M, but this isn’t much of an advantage, because we need to know M in order to interact with it at all, and so we might as well just call the static methods on the original type M directly. What we are wanting is a way to pass around a type-erased value that still gives us (type-erased) access to the underlying static interface of M but without instantiating a concrete M (i.e. no Box<dyn MyTrait>).

As usual, we use a trait TypeErasedTag to erase the concrete type M, and we provide a blanket implementation for all TypedTag<M: MyTrait>. But these aren’t values we can pass around directly. They need to be “behind” a reference or box. So to finish, we create a type alias Tag that represents a type-erased reference to a TypeErasedTag, and a factory function tag_for that builds the type-erased tag for a given concrete type M.

Structs

TypedTag

A ZST that “names” M. This gives us access to the concrete type M without instantiating it.

Traits

TypeErasedTag

As usual, we use a trait TypeErasedTag to erase the concrete type M, and we provide a blanket implementation for all TypedTag<M: MyTrait>.

Functions

tag_for

Factory: build the erased tag for a given M (without allocating). Compare to the equivalent function in simple_static_vtable. We require an additional 'static trait constraint, because Tag is a trait object rather than a concrete type.

Type Aliases

Tag

Type-erased handle: just a fat pointer (2 words), no alloc.