rust_patterns::type_erasure::static_interface::static_reference_to_zst

Function tag_for

pub fn tag_for<M: MyTrait + 'static>() -> Tag

Factory: build the erased tag for a given M (without allocating). Compare to the equivalent function in simple_static_vtable. We require an additional 'static trait constraint, because Tag is a trait object rather than a concrete type.