rust_patterns::data_structures

Module value_vector

ValueVec<T>: a by-value, ref-less vector with interior mutability.

You can convert to and from a Vec<T> at zero cost.

Key properties:

All mutating operations use &self (immutable receiver).

No references to elements are ever given out.

Elements are inserted/removed/moved by value.

Getting a value returns a Clone (or Copy) of the stored element.

Many shared immutable references to a ValueVec can exist simultaneously safely.

Trade-offs:

You can’t borrow into the backing memory; you pay cloning cost to read.

In return, the backing storage may reallocate at will without invalidating any external references (because none are ever issued).

Functionality:

For any functionality of Vec<T> that ValueVec<T> doesn’t provide, you can use into_vec to convert to a Vec<T> at zero cost.

Safety and Soundness:

ValueVec methods must not be re-entered while another method is still executing on the same instance. In particular, element types must not call back into the same ValueVec from within their Drop, Clone, or any other user code invoked during an operation on the ValueVec. Doing so would create overlapping borrows of the internal Vec and cause undefined behavior. You would have to try pretty hard to trigger this, but it is possible.

Structs

ValueVec

A by-value, ref-less vector with interior mutability. Values of type V can be moved into and out of the vector. Cloning / copying getters exist if V implements Clone/Copy