rust_patterns::plugins

Module item_registry

The Registry Pattern

The registry pattern is a plugin-like pattern for managing a collection of RegisteredItem instances such that discovery of items is automatic/implicit.

For our use case, the RegisteredItems are lazily registered (the allocation or instantiation of static data) and instantiated (for instance data), but this is not an essential part of the pattern. What is essential is that the RegisteredItems are automatically discovered, even if they are defined in external crates.

Note: If you are happy with the lazy discovery of RegisteredItems, then you can get away with a simpler implementation that uses neither “run before main()” nor “distributed slice” mechanisms (see ‘plugins’ module for discussion). There is a trade-off: You need a mechanism for interior mutability of RegisteredItems (the owner of the item instances) if you want to initialize the RegisteredItem on first access, because you don’t know how many RegisteredItem types there are globally when RegisteredItems is constructed and thus don’t know how big the vector should be.

In the simplest formulation, the “registry pattern” is just a vector of RegisteredItems, one instance for each concrete RegisteredItem type that exists, each one of which having a static index variable in which is stored its position in the vector that owns it. We just provide a mechanism for assigning the indices to the concrete RegisteredItem types at runtime in a logically consistent and thread-safe way that doesn’t require explicit knowledge of every RegisteredItem type that exists. To recap: We have one item instance per concrete RegisteredItem type per store.

In the next section we go into much more detail about the design decisions one has to make for any given implementation of the registry pattern. In that section we will point out which decisions we’ve made for the example we are illustrating in the code provided here, but here’s a brief rundown of the main points: We define the RegisteredItems struct as a wrapper type around the vector of RegisteredItem instances to provide convenient [RegisteredItems::get_item] and [RegisteredItems::get_item_mut] methods, which automatically instantiate the RegisteredItem instance if it has not yet been initialized. Note that we intentionally do not call the RegisteredItems struct a “registry” because it is not necessarily a singleton. In fact, we might have many RegisteredItems instances each holding RegisteredItem instances it owns. “The registry” that records which RegisteredItem types exist is purely abstract; depending on the underlying implementation, there might not be any actual list of which RegisteredItem types exist at all!

A RegisteredItem need not actually be a singleton, but it, or rather its type, is only ever registered once at runtime; that is, its index, which is shared among all instances of the type, is initialized only once (in a ctor). This allows a “registered item” type to have “singleton metadata” ¹ associated with it in addition to whatever instance data it needs. As already mentioned, at minimum each RegisteredItem knows its own index, which is used to access the item from the owning RegisteredItems container via simple indexing into a vector of RegisteredItems. (See the ‘plugins’ module for more details about implementation.)

Q: Wait, aren’t you just talking about static variables? Or, aren’t you just talking about lazy initialization of static data?

A: No, I am describing an interface for automatic discovery of special types and, (though only incidentally) the initialization of their static data. How rich the static data is depends entirely on your use case, so long as it at minimum has an index field to make accessing instances of the items in the owning container easy.

Notes

1. This is usually just called static data, or data in a static variable, but I am trying to avoid any assumptions about when this data is initialized and whether it is mutable.

Design Options for the Registry Pattern

In this section we explore in detail the different design choices one must make when implementing the registry pattern. We will call the RegisteredItem types of a particular implementation of the registry pattern the Widget types, or similar made up name. The owner will be some type wrapping a vector, which we’ll call the WidgetStore, or some variant of *Store.

Design Dimensions for Widgets and WidgetStores

Below is my attempt at carving up the design space for our registry pattern implementation with a few alternatives listed for each. I’m using the phrase dimension to refer to the major orthogonal design choices that need to be made about our container. By orthogonal I mean a choice in one dimension does not depend on a choice in another dimension. In actual practice, implementation choices in one dimension may affect implementation in another dimension, so this independence is approximate and “leaky”.

All registered items that participate in the same registry pattern implementation must share the semantics implied by these design choices to avoid significant complexity explosion.

Initialization time

In a singleton at program start (a la “distributed slice”, prototype style)

Upon WidgetStore creation (an instance for each Widget type that exists is eagerly created)

Lazily, e.g. upon first access in WidgetStore. The “slots” in the owning WidgetStore are occupied by OnceCells

Something else? E.g. super lazily: upon write access in WidgetStore.

In this example: We lazily instantiate Widgets upon first access. We use ctor magic to know how many “slots” the WidgetStore needs to allocate.

Ownership semantics: Specific to WidgetStore’s relationship to Widget values, not to views client code have into the data.

Widgets are owned by the WidgetStore which gives out regular mutable or immutable references

A WidgetStore is either literally or essentially a singleton (possibly a naked “distributed slice”), only one instance of each Widget type is created globally.

Widgets are shared, stored in Arc/Rc or Cow types.

Widgets are Copy types

Move semantics. I’m not sure what the use case would be, but it’s possible to move Widget instances out of the WidgetStore.

In this example: Widgets are owned by the WidgetStore which gives out regular mutable or immutable references

Interior mutability

“default” semantics: Widgets are owned by the WidgetStore, which is in turn owned by the Context (say), and they are all mutable or immutable at once accordingly.

WidgetStore lives in a RefCell, so entire store is mutable at once or immutable at once

WidgetStore holds individual Widgets in RefCells, so multiple distinct Widgets can be simultaneously mutated

Widgets are Copy types, so they can be copied around and mutated in place (possibly as atomic types)

Widgets implement their own interior mutability

In this example: “default” semantics (modulo OnceCell’s internal mechanisms); Mutation requires mutable access to the WidgetStore

Access control: what “views” into the data are enabled

Client code get’s “direct” access to a &Widget or &mut Widget

Client code passes a closure to a with_widget method, which closure is passed a &Widget; and similarly for mutable variant

WidgetStore provides rich API for interacting with Widgets, e.g. WidgetStore::froculate_widget::<Sprocket>() froculate’s a widget on the caller’s behalf. (We generally do this with DataPlugins, for example.)

Pass out Ref<'a, Widget>s / RefMut<'a, Widget>s to client code, which implement Deref/DerefMut to &Widget/&mut Widget

Widgets have Copy semantics, WidgetStore::get() returns a copy of the stored Widget. Have a corresponding WidgetStore::set() instead of WidgetStore::get_mut() with reference semantics.

YOLO raw pointers 🤣

In this example: Client code gets “direct” access to a &Widget or &mut Widget.

Other Dimensions: Some other dimensions I don’t have much to say about.

Type Discovery: We want “automatic” discovery, but some designs can get away with not using ctor/linkme tricks, in particular when discovery can happen on an as-needed (we “discover” ConcreteWidget only when WidgetStore::get::<ConcreteWidget>() is called); see notes below. This example uses ctor to know how many concrete Widget types there are globally and to assign each an index.

Concurrency Model: In this example, we take care to initialize each Widget type’s index in a thread-safe way but otherwise don’t share access across thread boundaries.

Lifetime Model: How long are Widget and WidgetStore instances expected to live? This implementation makes no assumptions.

“Leaky” Implementation Bits:

OnceCell has its own interior mutability mechanism; same with inventory, others; though they can be optimized for specific use cases (e.g. OnceCell is fast to access after initialization)

If WidgetStore does not need to be immutable, or if it or its wrapped vector can live in a RefCell, then magic ctor/linkme mechanisms are not necessary, because we don’t have to know the total count of all Widget types upon WidgetStore creation–we can grow the vector as needed.

If WidgetStore needs to be immutable, then we need to know the total count of all Widget types upon WidgetStore creation in order to allocate enough “slots” for all the Widget type even if we defer Widget creation by using OnceCells.

OnceCells push (a fast form of) interior mutability out from the Context or WidgetStore level to the individual Widget level

Eager initialization is an alternative to OnceCell and isn’t necessarily harder to implement, since you already pay the price of fancy linker tricks for the OnceCell solution.

Multiple registries for different thing categories vs. a single registry for all (Generic) Registered Items

We can imagine two different approaches to how many implementations of the registry pattern are in a given project:

A single implementation of the registry pattern could conceivably be generic over every category of thing that can be registered, or at least every category that matches all of your choices for you have made in all of the dimensions listed in the previous section.

Unifies the storage and access for Widgets, Gadgets, Blogets, Sprockets, …, potentially lots of disparate categories of objects.

Minimizes lines of code, API

You have separate implementations in your project of the registry pattern for each category even if the design choices in all dimensions of the previous section are the same. For example, you might implement the registry pattern for your Widget system and then have a separate independent implementation of the registry pattern for your Gadget system.

Different public API for different categories of objects

If we want to change the storage / ownership / access semantics for a category of objects, it’s already distinguished from other categories and has its own API.

My advice is to choose (2) almost always. Superficial implementation features are generally poor decomposition criteria. A Widget system and a Gadget system should generally not be artificially coupled.

Structs

RegisteredItems

A wrapper around a vector of RegisteredItems.

Statics

NEXT_ITEM_INDEX

Global item index counter, keeps track of the index that will be assigned to the next item that requests an index. Equivalently, holds a count of the number of items currently registered.

REGISTERED_ITEMS

For simple applications, just knowing how many items are registered is enough, and that information is already captured in NEXT_ITEM_INDEX. In more sophisticated implementations, we would want to store metadata about the items that are registered.

Traits

RegisteredItem

A trait for items that can be registered (DataPlugin, PersonProperty)

Functions

add_to_registry

Adds a new item to the registry. The job of this method is to create whatever “singleton” data/metadata is associated with the RegisteredItem if it doesn’t already exist. The “registry” is a global singleton (literally just NEXT_ITEM_INDEX in this implementation), so it is enough to know the concrete type of the RegisteredItem. For this illustration, the only singleton metadata that is instantiated is the item’s index, which we return in case the caller wants it.

get_registered_item_count

An accessor for NEXT_ITEM_INDEX

initialize_item_index

Encapsulates the synchronization logic for initializing an item’s index.

Type Aliases

BxRegisteredItem

A boxed RegisteredItem